Rename/move watch.Batcher to sync.WriteBatcher
